pyproject.toml
[tool.poetry]
name = "sahabot2"
version = "0.1.0"
description = "SahasrahBot2 - A NiceGUI + Tortoise ORM web application with Discord OAuth2"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
nicegui = "^3.2.0"
tortoise-orm = "^0.25.0"
aerich = "^0.9.2"
python-dotenv = "^1.0.0"
httpx = "^0.26.0"
pydantic = "^2.12.0"
pydantic-settings = "^2.11.0"
"discord.py" = "^2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
black = "^24.0.0"
ruff = "^0.1.0"

[tool.aerich]
tortoise_orm = "migrations.tortoise_config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
